pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "webnovels"
version = "0.0.1"
requires-python = ">= 3.11"
dependencies = [
    "pypub3",
    "requests",
    "bs4",
    "selenium",
    "pyspellchecker",
    "python-docx",
]

[tool.setuptools]
package-dir = {"" = "src"}